Created TablePluginsChecker, CollectionPluginsChecker
